Wrapping up initial work on Commitment
Its an odd bit of work. The function of this gem is to inject behavior
into the build of a Rails project. It has been tested, by hand, on the
cogitate application (https://github.com/ndlib/cogitate). However it
does not have what I consider to be the customary tests to help assure
adopters that things are working.